feat: refactor layout component
