+++ frontend/src/components/layout/MinimalLayout.tsx
