+++ frontend/src/components/layout/part/SidebarContent.tsx
